drtmle

Main function to estimate TMLE with doubly-robust inference

drtmle(Y, A, W, a0 = unique(A), family = "binomial", stratify = TRUE,
  libraryQ = NULL, libraryg = NULL, libraryQr = NULL, librarygr = NULL,
  glmQ = NULL, glmg = NULL, glmQr = NULL, glmgr = NULL, guard = c("Q",
  "g"), reduction = "univariate", returnModels = FALSE, maxIter = 100,
  tolIC = 1/length(Y), tolg = 1e-08, verbose = TRUE, Qsteps = 2, ...)

Arguments

Y
A vector of continuous or binary outcomes.

A
A vector of binary treatment assignment (assumed to be equal to 0 or 1)

W
A data.frame of named covariates

a0
A vector of treatment levels at which to compute the adjusted mean outcome.

family
A character equal to either "binomial" or "gaussian". This is passed to the SuperLearner function internally.

stratify
A boolean indicating whether to estimate the outcome regression separately for observations with A equal to 0/1 (if TRUE) or to pool across A (if FALSE).

libraryQ
A vector of characters or a list describing the Super Learner library to be used for the outcome regression. See link{SuperLearner::SuperLearner} for details.

libraryg
A vector of characters or a list describing the Super Learner library to be used for the propensity score. See link{SuperLearner::SuperLearner} for details.

libraryQr
A vector of characters or a list describing the Super Learner library to be used for the first reduced-dimension regression.

librarygr
A vector of characters or a list describing the Super Learner library to be used for the second reduced-dimension regression.

glmQ
A character describing a formula to be used in the call to glm for the outcome regression. Ignored if libraryQ!=NULL.

glmg
A character describing a formula to be used in the call to glm for the propensity score. Ignored if libraryg!=NULL.

glmQr
A character describing a formula to be used in the call to glm for the first reduced-dimension regression. Ignored if libraryQr!=NULL.

glmgr
A character describing a formula to be used in the call to glm for the second reduced-dimension regression. Ignored if librarygr!=NULL.

guard
A character vector indicating what pattern of misspecifications to guard against. If guard contains "Q", then the TMLE guards against misspecification of the outcome regression. If guard contains "g" then the TMLE (additionally) guards against misspecification of the propensity score. If NULL, the usual TMLE is computed.

reduction
A character equal to "univariate" for a univariate misspecification correction or "bivariate" for the bivariate version.

returnModels
A boolean indicating whether to return model fits for the outcome regression, propensity score, and reduced-dimension regressions.

maxIter
A numeric that sets the maximum number of iterations the TMLE can perform in its fluctuation step.

tolIC
A numeric that defines the stopping criteria based on the empirical mean of the scores of the fluctuation submodels submodels. Setting to "default"

tolg
A numeric indicating the minimum value for estimates of the propensity score.

verbose
A boolean indicating whether to print status updates.

Qsteps
A numeric equal to 1/2 indicating whether the fluctuation submodel for the outcome regression should be fit using a single minimization (Qsteps = 1) or a backfitting-type minimization (Qsteps=2). The latter was found to be more stable in simulations.

...
Other options (not currently used)