estimateQrn

Estimates the reduced dimension regressions necessary for the fluctuations of g

estimateQrn(Y, A, W, DeltaA, DeltaY, Qn, gn, glm_Qr, SL_Qr,
  family = stats::gaussian(), a_0, returnModels, validRows = NULL)

Arguments

Y
A vector of continuous or binary outcomes.

A
A vector of binary treatment assignment (assumed to be equal to 0 or 1)

W
A data.frame of named covariates

DeltaA
Indicator of missing treatment (assumed to be equal to 0 if missing 1 if observed)

DeltaY
Indicator of missing outcome (assumed to be equal to 0 if missing 1 if observed)

Qn
A list of outcome regression estimates evaluated on observed data. If NULL then 0 is used for all Qn (as is needed to estimate reduced dimension regression for adaptive_iptw)

gn
A list of propensity regression estimates evaluated on observed data

glm_Qr
A character describing a formula to be used in the call to glm for the first reduced-dimension regression. Ignored if SL_gr!=NULL.

SL_Qr
A vector of characters or a list describing the Super Learner library to be used for the first reduced-dimension regression.

family
Should be gaussian() unless called from adaptive_iptw with binary Y.

a_0
A list of fixed treatment values.

returnModels
A boolean indicating whether to return model fits for the outcome regression, propensity score, and reduced-dimension regressions.

validRows
A list of length cvFolds containing the row indexes of observations to include in validation fold.