estimategrn

Estimates the reduced dimension regressions necessary for the additional fluctuations.

estimategrn(Y, A, W, Qn, gn, librarygr, tolg, glmgr, a0, reduction)

Arguments

Y
A vector of continuous or binary outcomes.

A
A vector of binary treatment assignment (assumed to be equal to 0 or 1)

W
A data.frame of named covariates

Qn
A list of outcome regression estimates evaluated on observed data

gn
A list of propensity regression estimates evaluated on observed data

librarygr
A vector of characters or a list describing the Super Learner library to be used for the second reduced-dimension regression.

tolg
A numeric indicating the minimum value for estimates of the propensity score.

glmgr
A character describing a formula to be used in the call to glm for the second reduced-dimension regression. Ignored if librarygr!=NULL.

a0
A list of fixed treatment values

reduction
A character equal to "univariate" for a univariate misspecification correction or "bivariate" for the bivariate version.