fluctuateQ2

Function called internally by drtmle to perform the second fluctuation of the initial estimator of Q (i.e., solves the new estimating eqn that results from misspecification of g)

fluctuateQ2(Y, A, W, DeltaY, DeltaA, Qn, gn, grn, a_0, reduction,
  coefTol = 1e+05)

Arguments

Y
The outcome

A
The treatment

W
The covariates

DeltaY
Indicator of missing outcome (assumed to be equal to 0 if missing 1 if observed)

DeltaA
Indicator of missing treatment (assumed to be equal to 0 if missing 1 if observed)

Qn
A list of outcome regression estimates evaluated on observed data

gn
A list of propensity regression estimates evaluated on observed data

grn
A list of reduced-dimension regression estimates evaluated on observed data

a_0
A list of fixed treatment values

reduction
A character indicating what reduced dimension regression was used.

coefTol
A tolerance level on the magnitude of the coefficient that flags the result as potentially the result of numeric instability.