TMLE estimate of the average treatment effect with doubly-robust inference

TMLE estimate of the average treatment effect with doubly-robust inference

drtmle(Y, A, W, DeltaA = as.numeric(!is.na(A)),
  DeltaY = as.numeric(!is.na(Y)), a_0 = unique(A[!is.na(A)]),
  family = stats::binomial(), stratify = TRUE, SL_Q = NULL, SL_g = NULL,
  SL_Qr = NULL, SL_gr = NULL, glm_Q = NULL, glm_g = NULL,
  glm_Qr = NULL, glm_gr = NULL, guard = c("Q", "g"),
  reduction = "univariate", returnModels = FALSE, cvFolds = 1,
  maxIter = 3, tolIC = 1/length(Y), tolg = 0.01, verbose = FALSE,
  Qsteps = 2, parallel = FALSE, ...)

Arguments

Y
numeric of continuous or binary outcomes.

A
A vector of discrete-valued treatment assignment (assumed to be equal to 0 or 1).

W
A data.frame of named covariates.

DeltaA
A vector of missing treatment indicator (assumed to be equal to 0 if missing 1 if observed).

DeltaY
A vector of missing outcome indicator (assumed to be equal to 0 if missing 1 if observed).

a_0
A vector of treatment levels at which to compute the adjusted mean outcome (should correspond to values of A).

family
A family object equal to either binomial() or gaussian(), to be passed to the SuperLearner or glm function.

stratify
A boolean indicating whether to estimate the outcome regression separately for different values of A (if TRUE) or to pool across A (if FALSE).

SL_Q
A vector of characters or a list describing the Super Learner library to be used for the outcome regression. See link{SuperLearner::SuperLearner} for details.

SL_g
A vector of characters or a list describing the Super Learner library to be used for the propensity score. See link{SuperLearner::SuperLearner} for details.

SL_Qr
A vector of characters or a list describing the Super Learner library to be used for the reduced-dimension outcome regression.

SL_gr
A vector of characters or a list describing the Super Learner library to be used for the second reduced-dimension propensity score.

glm_Q
A character describing a formula to be used in the call to glm for the outcome regression. Ignored if SL_Q!=NULL.

glm_g
A character describing a formula to be used in the call to glm for the propensity score. Ignored if SL_g!=NULL.

glm_Qr
A character describing a formula to be used in the call to glm for reduced-dimension outcome regression. Ignored if SL_Qr!=NULL. The formula should use the variable name 'gn'.

glm_gr
A character describing a formula to be used in the call to glm for the reduced-dimension propensity score. Ignored if SL_gr!=NULL. The formula should use the variable name 'Qn' and 'gn' if reduction='bivariate' and 'Qn' otherwise.

guard
A character vector indicating what pattern of misspecifications to guard against. If guard contains "Q", then the TMLE guards against misspecification of the outcome regression by estimating the reduced-dimension outcome regression specified by glm_Qr or SL_Qr. If guard contains "g" then the TMLE (additionally) guards against misspecification of the propensity score by estimating the reduced-dimension propensity score specified by glm_gr or SL_gr.

reduction
A character equal to "univariate" for a univariate misspecification correction (default) or "bivariate" for the bivariate version.

returnModels
A boolean indicating whether to return model fits for the outcome regression, propensity score, and reduced-dimension regressions.

cvFolds
A numeric equal to the number of folds to be used in cross-validated fitting of nuisance parameters. If cvFolds = 1, no cross-validation is used.

maxIter
A numeric that sets the maximum number of iterations the TMLE can perform in its fluctuation step.

tolIC
A numeric that defines the stopping criteria based on the empirical mean of the scores of the fluctuation submodels submodels.

tolg
A numeric indicating the minimum value for estimates of the propensity score.

verbose
A boolean indicating whether to print status updates.

Qsteps
A numeric equal to 1 or 2 indicating whether the fluctuation submodel for the outcome regression should be fit using a single minimization (Qsteps = 1) or a backfitting-type minimization (Qsteps=2). The latter was found to be more stable in simulations and is the default.

parallel
A boolean indicating whether to use foreach to estimate nuisance parameters in parallel. Only useful if there is a registered parallel backend and cvFolds > 1.

...
Other options (not currently used).

Value

An object of class "drtmle".

drtmle

A list of doubly-robust point estimates and a doubly-robust covariance matrix

nuisance_drtmle

A list of the final TMLE estimates of the outcome regression ($QnStar), propensity score ($gnStar), and reduced-dimension regressions ($QrnStar, $grnStar) evaluated at the observed data values.

ic_drtmle

A list of the empirical mean of the efficient influence function ($eif) and the extra pieces of the influence function resulting from misspecification. All should be smaller than tolIC (unless maxIter was reached first).

aiptw_c

A list of doubly-robust point estimates and a non-doubly-robust covariance matrix. Theory does not guarantee performance of inference for these estimators, but simulation studies showed they often perform adequately.

nuisance_aiptw

A list of the initial estimates of the outcome regression, propensity score, and reduced-dimension regressions evaluated at the observed data values.

tmle

A list of doubly-robust point estimates and non-doubly-robust covariance for the standard TMLE estimator.

aiptw

A list of doubly-robust point estimates and non-doubly-robust covariance matrix for the standard AIPTW estimator.

gcomp

A list of non-doubly-robust point estimates and non-doubly-robust covariance matrix for the standard G-computation estimator. If super learner is used there is no guarantee of correct inference for this estimator.

QnMod

The fitted object for the outcome regression. Returns NULL if returnModels = FALSE.

gnMod

The fitted object for the propensity score. Returns NULL if returnModels = FALSE.

QrnMod

The fitted object for the reduced-dimension regression that guards against misspecification of the outcome regression. Returns NULL if returnModels = FALSE.

grnMod

The fitted object for the reduced-dimension regression that guards against misspecification of the propensity score. Returns NULL if returnModels = FALSE.

a_0

The treatment levels that were requested for computation of covariate-adjusted means.

Examples

# load super learner
library(SuperLearner)
# simulate data
set.seed(123456)
n <- 200
W <- data.frame(W1 = runif(n), W2 = rnorm(n))
A <- rbinom(n,1,plogis(W$W1 - W$W2))
Y <- rbinom(n, 1, plogis(W$W1*W$W2*A))
# fit drtmle
fit1 <- drtmle(W = W, A = A, Y = Y, a_0 = c(1,0),
               family=binomial(),
               stratify=FALSE,
               SL_Q=c("SL.glm","SL.mean","SL.step"),
               SL_g=c("SL.glm","SL.mean","SL.step"),
               SL_Qr="SL.npreg",
               SL_gr="SL.npreg")